// src/types/Palette.ts
/**
 * Colour palette as defined in UX terms
 * TODO: complete with the accent colours and others
 */
export type Palette = {
	brandPrimary: string;
	brand30: string;
	brand10: string;

	primary100: string;
	primary90: string;
	primaryPrimary: string;
	primary60: string;
	primary40: string;
	primary20: string;
	primary10: string;

	secondaryPrimary: string;
	secondary40: string;
	secondary20: string;
	secondary10: string;

	darkBackground100: string;
	darkBackgroundPrimary: string;
	darkBackground80: string;
	darkBackground30: string;

	lightBackgroundPrimary: string;

	primaryTextPrimary: string;

	secondaryTextPrimary: string;
	secondaryText50: string;

	white: string;

	info100: string;
	infoPrimary: string;
	info70: string;
	info30: string;
	info20: string;
	info10: string;

	error100: string;
	errorPrimary: string;
	error70: string;
	error30: string;
	error20: string;
	error10: string;

	warning100: string;
	warningPrimary: string;
	warning70: string;
	warning30: string;
	warning20: string;
	warning10: string;

	success100: string;
	successPrimary: string;
	success70: string;
	success30: string;
	success20: string;
	success10: string;
};